Added the countries data

+++ app/src/main/java/com/livedata/covid19/vo/CountryInfo.kt
@@ -9,6 +9,6 @@ data class CountryInfo(
     val id: Int,
     val iso2: String,
     val iso3: String,
-    val lat: Int,
-    val long: Int
+    val lat: Float,
+    val long: Float
 )
